backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pymongo==4.8.0
motor==3.5.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
numpy==2.0.0
scikit-learn==1.5.0
python-dotenv==1.0.1
httpx==0.27.0
